Get/set time for submission as java.util.Date
